+++ server/src/main/java/com/linbit/linstor/api/SpaceInfo.java
@@ -1,5 +1,8 @@
 package com.linbit.linstor.api;
 
+import com.linbit.linstor.core.objects.StorPool;
+import com.linbit.linstor.storage.StorageException;
+
 public class SpaceInfo
 {
     public Long totalCapacity;
@@ -9,5 +12,17 @@ public class SpaceInfo
     {
         totalCapacity = totalCapacityRef;
         freeCapacity = freeSpaceRef;
+    }
+
+    public static SpaceInfo buildOrThrowOnError(Long totalCapacityRef, Long freeSpaceRef, StorPool spRef)
+        throws StorageException
+    {
+        if (totalCapacityRef == null || freeSpaceRef == null)
+        {
+            throw new StorageException(
+                "Failed to query total capacity and/or free space for storage pool: " + spRef.getName()
+            );
+        }
+        return new SpaceInfo(totalCapacityRef, freeSpaceRef);
     }
 }